+++ src/Validators/ValidatorFactory.php
@@ -88,7 +88,7 @@ class ValidatorFactory implements ValidatorFactoryInterface
             $expectedType,
             $expectedId,
             $attributes,
-            $relationships,
+            $relationships ?: $this->relationships(),
             $context
         );
     }
@@ -102,13 +102,24 @@ class ValidatorFactory implements ValidatorFactoryInterface
     }
 
     /**
+     * @inheritDoc
+     */
+    public function relationship($expectedType = null, $allowEmpty = true, $acceptable = null)
+    {
+        return new RelationshipValidator(
+            $this->validationErrors,
+            $this->store,
+            $expectedType,
+            $allowEmpty,
+            $acceptable
+        );
+    }
+
+    /**
      * @inheritdoc
      */
-    public function hasOne(
-        $expectedType,
-        $allowEmpty = true,
-        $acceptable = null
-    ) {
+    public function hasOne($expectedType, $allowEmpty = true, $acceptable = null)
+    {
         return new HasOneValidator(
             $this->validationErrors,
             $this->store,
@@ -121,11 +132,8 @@ class ValidatorFactory implements ValidatorFactoryInterface
     /**
      * @inheritdoc
      */
-    public function hasMany(
-        $expectedType,
-        $allowEmpty = false,
-        $acceptable = null
-    ) {
+    public function hasMany($expectedType, $allowEmpty = false, $acceptable = null)
+    {
         return new HasManyValidator(
             $this->validationErrors,
             $this->store,